Revert "Various things related to TF2 SDK"
This reverts commit 60cb4ffc081bc3c14ba1f40f210ce3b64e63992d.

+++ src/materialsystem/shaderapidx9/locald3dtypes.h
@@ -12,8 +12,6 @@
 #pragma once
 #endif
 
-#include "tier0/platform.h"
-
 #if defined( DX10 ) && !defined( DX_TO_GL_ABSTRACTION )
 
 #include <d3d10.h>
@@ -115,8 +113,8 @@ typedef void *HardwareShader_t;
 //-----------------------------------------------------------------------------
 // The vertex and pixel shader type
 //-----------------------------------------------------------------------------
-typedef intp VertexShader_t;
-typedef intp PixelShader_t;	
+typedef int VertexShader_t;
+typedef int PixelShader_t;	
 
 //-----------------------------------------------------------------------------
 // Bitpattern for an invalid shader
